Fixed bugs in testing of journalling (AbstractDiagram -> String) and consequently fixed bugs in journalling.
Weakened test checksum to ignore the m_id of CurveLabels.  Two diagrams
which have a splilt contour now have the same checksum if they look the
same, and a re-ordering of m_ids will not break tests.  

NB Journalling effectively re-orders the zone descriptors in the diagram
description, so checksums which are order-sensitive will trigger test
failures when we switch test_journalling to true.  At the moment, all
tests pass if we test_journalling.

New test number 133 is a problem case for layouts - a spider foot is
coincident with the leg of another spider.